refactor : Context 파일 생성하여 add, remove 함수 관리
